requirements.txt
# requirements.txt (versão final)
pg8000
PyJWT
boto3
python-dotenv